miniCPM-o_2.6_ipex_llm/requirements_o2.6_ipex.txt
Pillow==10.1.0
transformers==4.44.2
sentencepiece==0.2.0
vector-quantize-pytorch==1.18.5
vocos==0.1.0
accelerate==1.2.1
timm==0.9.10
soundfile==0.12.1
librosa==0.9.0
decord
moviepy

# for web demo
aiofiles==23.2.1
onnxruntime==1.20.1
fastapi
uvicorn
streamlit

#gradio==4.44.1
#http://thunlp.oss-cn-qingdao.aliyuncs.com/multi_modal/never_delete/modelscope_studio-0.4.0.9-py3-none-any.whl

